update of the console to allow all commands to work with all classes created previously.

+++ console.py
@@ -3,6 +3,12 @@
 
 import cmd
 from models.base_model import BaseModel
+from models.user import User
+from models.state import State
+from models.city import City
+from models.amenity import Amenity
+from models.place import Place
+from models.review import Review
 from models import storage
 
 
@@ -31,6 +37,28 @@ class HBNBCommand(cmd.Cmd):
         else:
             if arg == "BaseModel":
                 my_model = BaseModel()
+                exist = 1
+            elif arg == "User":
+                my_model = User()
+                exist = 1
+            elif arg == "State":
+                my_model = State()
+                exist = 1
+            elif arg == "City":
+                my_model = City()
+                exist = 1
+            elif arg == "Amenity":
+                my_model = Amenity()
+                exist = 1
+            elif arg == "Place":
+                my_model = Place()
+                exist = 1
+            elif arg == "Review":
+                my_model = Review()
+                exist = 1
+            else:
+                exist = 0
+            if exist == 1:
                 my_model.save()
                 print("{}".format(my_model.id))
             else:
@@ -45,7 +73,11 @@ class HBNBCommand(cmd.Cmd):
             print("** class name missing **")
         else:
             args = arg.split(" ")
-            if args[0] == "BaseModel":
+            listclass = ['BaseModel', 'User', 'City', 'State']
+            listclass.append('Amenity')
+            listclass.append('Place')
+            listclass.append('Review')
+            if args[0] in listclass:
                 if len(args) < 2:
                     print("** instance id missing **")
                 else:
@@ -56,10 +88,11 @@ class HBNBCommand(cmd.Cmd):
                     for obj_id in all_objs.keys():
                         obj_split = obj_id.split(".")
                         if obj_split[1] == id_instance:
-                            ob = all_objs[obj_id]
-                            print(ob)
-                            find = 1
-                            break
+                            if obj_split[0] == args[0]:
+                                ob = all_objs[obj_id]
+                                print(ob)
+                                find = 1
+                                break
                     if find == 0:
                         print("** no instance found **")
             else:
@@ -74,7 +107,11 @@ class HBNBCommand(cmd.Cmd):
             print("** class name missing **")
         else:
             args = arg.split(" ")
-            if args[0] == "BaseModel":
+            listclass = ['BaseModel', 'User', 'City', 'State']
+            listclass.append('Amenity')
+            listclass.append('Place')
+            listclass.append('Review')
+            if args[0] in listclass:
                 if len(args) < 2:
                     print("** instance id missing **")
                 else:
@@ -85,11 +122,12 @@ class HBNBCommand(cmd.Cmd):
                     for obj_id in all_objs.keys():
                         obj_split = obj_id.split(".")
                         if obj_split[1] == id_instance:
-                            ob = all_objs[obj_id]
-                            storage.destroy_objects(obj_id)
-                            storage.save()
-                            find = 1
-                            break
+                            if obj_split[0] == args[0]:
+                                ob = all_objs[obj_id]
+                                storage.destroy_objects(obj_id)
+                                storage.save()
+                                find = 1
+                                break
                     if find == 0:
                         print("** no instance found **")
             else:
@@ -105,7 +143,11 @@ class HBNBCommand(cmd.Cmd):
             for obj_id in all_objs.keys():
                 print(all_objs[obj_id])
         else:
-            if arg == "BaseModel":
+            listclass = ['BaseModel', 'User', 'City', 'State']
+            listclass.append('Amenity')
+            listclass.append('Place')
+            listclass.append('Review')
+            if arg in listclass:
                 for obj_id in all_objs.keys():
                     obj_split = obj_id.split(".")
                     if obj_split[0] == arg:
@@ -123,7 +165,11 @@ of all instances based or not on the class name')
             print("** class name missing **")
         else:
             args = arg.split(" ")
-            if args[0] == "BaseModel":
+            listclass = ['BaseModel', 'User', 'City', 'State']
+            listclass.append('Amenity')
+            listclass.append('Place')
+            listclass.append('Review')
+            if args[0] in listclass:
                 if len(args) < 4:
                     if len(args) == 1:
                         print("** instance id missing **")
@@ -139,10 +185,13 @@ of all instances based or not on the class name')
                     for obj_id in all_objs.keys():
                         obj_split = obj_id.split(".")
                         if obj_split[1] == id_instance:
-                            storage.update_objects(obj_id, args[2], args[3])
-                            storage.save()
-                            find = 1
-                            break
+                            if obj_split[0] == args[0]:
+                                a1 = args[2]
+                                a2 = args[3]
+                                storage.update_objects(obj_id, a1, a2)
+                                storage.save()
+                                find = 1
+                                break
                     if find == 0:
                         print("** no instance found **")
             else:
@@ -155,4 +204,8 @@ by adding or updating attribute')
 
 
 if __name__ == '__main__':
-    HBNBCommand().cmdloop()
+    import sys
+    if len(sys.argv) > 1:
+        HBNBCommand().onecmd(' '.join(sys.argv[1:]))
+    else:
+        HBNBCommand().cmdloop()
